services/backtesting/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grid-backtester"
version = "1.0.0"
description = "Standalone grid backtesting and optimization service"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.1",
    "numpy>=1.26",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "structlog>=23.2",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.25",
    "aiosqlite>=0.19",
    "plotly>=5.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "httpx>=0.26",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
